Bug fixing default value behaviour on track widget and quick track

+++ app/src/main/java/com/samco/trackandgraph/displaytrackgroup/DataPointInputView.kt
@@ -102,7 +102,7 @@ class DataPointInputView(context: Context, private val state: DataPointInputData
                 true
             } else false
         }
-        val text = if (state.value == 0.toDouble()) "" else doubleFormatter.format(state.value)
+        val text = if (state.value == 0.0) "" else doubleFormatter.format(state.value)
         numberInput.setText(text)
     }
